added delete s3 asset func

+++ pages/api/file/delete/[id].ts
@@ -1,5 +1,7 @@
+import { File } from '@prisma/client';
 import prisma from 'lib/prisma';
 import type { NextApiRequest, NextApiResponse } from 'next';
+import { deleteFile } from 'services/uploadService';
 
 export default async function handle(req: NextApiRequest, res: NextApiResponse) {
   if (req.method !== 'DELETE') {
@@ -10,6 +12,13 @@ export default async function handle(req: NextApiRequest, res: NextApiResponse) 
     return res.status(400).json({ message: 'Invalid request' });
   }
   try {
+    const file: File = await prisma.File.findUnique({
+      where: { Id: Number(id) },
+    });
+    if (!file) {
+      return res.status(404).json({ message: 'File not found' });
+    }
+    await deleteFile(file.Location);
     const deletedFile = await prisma.File.delete({
       where: { Id: Number(id) },
     });
